fix AIP rendering issue
