+++ .changeset/orange-icons-battle.md
@@ -1,0 +1,5 @@
+---
+"@osdk/create-app": patch
+---
+
+update AIP template
